backend/requirements.txt
APScheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
statsmodels>=0.14,<1.0
prophet>=1.1,<2.0
plotly>=5.0,<6.0
